LayoutTests/js/function-declaration-statement-expected.txt

This test checks that function declarations are treated as statements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ifTest() is true
PASS ifElseTest() is true
PASS labelTest() is true
PASS successfullyParsed is true

TEST COMPLETE
